updated css with float rule and border-radius
updated css with float rule for the the two items i needed at the bottom. Called on toe, and the other nail because I wasn't sure what to call them. Floated one left, and the other right. Will need to link them to navigate to the correct sections of the page.