updated float rules
Needed to update float rules to make sure my layout acts correctly. Due to that, I removed most of the floats, for the h2 and paragraph tags to format everything correctly. In addition, because I updated the order of the files, I found that I accidentally forgot to add a width property to the .calculator class.