updated styles.css and index.html
updated styles.css by removing the .main-nav class and exchanging it with the class nav. Also made sure to fix the nav menu. Floated the parent nav element in the header, and the floated the list items to the left of its parent element to make sure everything worked correctly.

For index.html, I updated the .main-nav menu with the  semantic nav class in order to get the exceeds expectations grade.